feat:footer 수정
