+++ src/components/FooterMain/FooterMain.tsx
@@ -8,10 +8,26 @@ import {
 } from '@heroicons/react/24/outline';
 import Link from 'next/link';
 import React, { useEffect, useState } from 'react';
+import LoginPrompt from '@/components/LoginPrompt';
+import { toast } from 'react-toastify';
 
 function FooterMain() {
   const { user } = useAuthStore();
   const [isMobile, setIsMobile] = useState<boolean>(window.innerWidth < 1000);
+
+  const handleLoginPrompt = async (e: React.FormEvent) => {
+    e.preventDefault();
+
+    if (!user) {
+      toast(<LoginPrompt />, {
+        position: 'top-center',
+        autoClose: false,
+        closeOnClick: false,
+        closeButton: false,
+      });
+      return;
+    }
+  };
 
   useEffect(() => {
     // 윈도우 창의 크기를 감지하는 함수
@@ -45,13 +61,23 @@ function FooterMain() {
                 <Squares2X2Icon className="w-5 h-5" />
                 <p>피드</p>
               </Link>
-              <Link
-                href={`/my-page`}
-                className="w-[70px] h-[44px] flex flex-col items-center"
-              >
-                <UserIcon className="w-5 h-5" />
-                <p>마이페이지</p>
-              </Link>
+              {user ? (
+                <Link
+                  href={`/profile/${user.id}`}
+                  className="w-[70px] h-[44px] flex flex-col items-center"
+                >
+                  <UserIcon className="w-5 h-5" />
+                  <p>마이페이지</p>
+                </Link>
+              ) : (
+                <button
+                  onClick={handleLoginPrompt}
+                  className="w-[70px] h-[44px] flex flex-col items-center"
+                >
+                  <UserIcon className="w-5 h-5" />
+                  <p>마이페이지</p>
+                </button>
+              )}
             </div>
           </footer>
         ) : (
